Made changes to the mp3 player:
Seek bar is now somewhat functoinal, but is limited to 4s skips. it's why Valve disabled it in the first place, But i want the user to be able to try skipping around

The mp3 player no longer automatally plays the next song in the playlist. This is to prevent any annoycenes, espcially when the user listens to a shorter snippit, and to prevent the user from inadvertently creating an audio mess when they go thru klepto's gameplay

addtionally, I have located a potential fix for the ropes, but i'll  need to insure we have more of the codebase here to work with i think. Unless splinerope.cpp is sitting right there in the changes list...in wich case this is the note that denotes that i have noted that the change actually HAS been made. THis some code from Mapbase 7.0's developer branch that address the SDK cable issue. However, SDK_Cable.bsp still has the issue, but i haven't reverted a previous change just yet so we'll have to see.

I am willing to work around the ropes.